Added initial interleaving algorithm.
In the case of 2 empty cells, choose A if size of set A is larger than Set B,
else choose B.

If one cell is empty and the other is not, get cost of placing non-empty cell
in the current site and next site. If next site has less cost, then place empty
cell in the current site.